+++ android/src/main/java/com/example/r_upgrade/common/tasks/GenerateAndInstallAsyncTask.java
@@ -2,16 +2,16 @@ package com.example.r_upgrade.common.tasks;
 
 import android.app.DownloadManager;
 import android.content.Context;
-import android.content.Intent;
 import android.database.Cursor;
 import android.net.Uri;
 import android.os.AsyncTask;
 import android.os.Build;
-import android.util.Log;
 
 import com.example.r_upgrade.RUpgradeFileProvider;
-import com.example.r_upgrade.common.HotUpgradeManager;
-import com.example.r_upgrade.common.IncrementUpgradeManager;
+import com.example.r_upgrade.common.install.BaseInstallFactory;
+import com.example.r_upgrade.common.install.SilentInstallFactory;
+import com.example.r_upgrade.common.manager.HotUpgradeManager;
+import com.example.r_upgrade.common.manager.IncrementUpgradeManager;
 import com.example.r_upgrade.common.RUpgradeLogger;
 import com.example.r_upgrade.common.UpgradeSQLite;
 
@@ -28,11 +28,14 @@ public class GenerateAndInstallAsyncTask extends AsyncTask<Integer, Integer, Uri
     final WeakReference<Context> contextWrapper;
     boolean isUserDownloadManager;
     MethodChannel.Result result;
+    private BaseInstallFactory installFactory;
 
-    public GenerateAndInstallAsyncTask(Context context, boolean isUserDownloadManager, MethodChannel.Result result) {
+    public GenerateAndInstallAsyncTask(Context context, boolean isUserDownloadManager, MethodChannel.Result result,
+                                       BaseInstallFactory installFactory) {
         this.contextWrapper = new WeakReference<Context>(context);
         this.isUserDownloadManager = isUserDownloadManager;
         this.result = result;
+        this.installFactory = installFactory;
     }
 
     @Override
@@ -63,6 +66,9 @@ public class GenerateAndInstallAsyncTask extends AsyncTask<Integer, Integer, Uri
                 } else {
                     uri = Uri.fromFile(file);
                 }
+                if (installFactory != null && installFactory instanceof SilentInstallFactory) {
+                    uri = Uri.parse(file.getPath());
+                }
                 if (upgradeFlavor == UpgradeSQLite.UPGRADE_FLAVOR_INCREMENT) {
                     String newPath = new IncrementUpgradeManager(contextWrapper.get()).mixinAndGetNewApk(path);
                     RUpgradeLogger.get().d(TAG, "合成成功，新的安装包路径：" + newPath);
@@ -72,6 +78,9 @@ public class GenerateAndInstallAsyncTask extends AsyncTask<Integer, Integer, Uri
                         uri = RUpgradeFileProvider.getUriForFile(contextWrapper.get(), contextWrapper.get().getApplicationInfo().packageName + ".fileProvider", file);
                     } else {
                         uri = Uri.fromFile(file);
+                    }
+                    if (installFactory != null && installFactory instanceof SilentInstallFactory) {
+                        uri = Uri.parse(file.getPath());
                     }
                 } else if (upgradeFlavor == UpgradeSQLite.UPGRADE_FLAVOR_HOT_UPDATE) {
                     boolean isSuccess = new HotUpgradeManager(contextWrapper.get()).hotUpgrade(uri);
@@ -85,6 +94,7 @@ public class GenerateAndInstallAsyncTask extends AsyncTask<Integer, Integer, Uri
         } catch (Exception e) {
             e.printStackTrace();
         }
+
         return uri;
     }
 
@@ -117,20 +127,7 @@ public class GenerateAndInstallAsyncTask extends AsyncTask<Integer, Integer, Uri
 
     //安装apk
     private boolean installApk(Uri uri) {
-        Intent install = new Intent(Intent.ACTION_VIEW);
-        if (uri != null) {
-            RUpgradeLogger.get().d(TAG, uri.toString());
-            install.setDataAndType(uri, "application/vnd.android.package-archive");
-            install.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
-                install.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
-            } else {
-                install.addCategory(Intent.CATEGORY_DEFAULT);
-            }
-            contextWrapper.get().startActivity(install);
-            return true;
-        } else {
-            return false;
-        }
+        if (installFactory == null) return true;
+        return installFactory.install(uri);
     }
 }