+++ android/src/main/java/com/example/r_upgrade/common/tasks/GenerateAndInstallByPathAsyncTask.java
@@ -2,20 +2,20 @@ package com.example.r_upgrade.common.tasks;
 
 import android.content.Context;
 import android.content.Intent;
-import android.database.Cursor;
 import android.net.Uri;
 import android.os.AsyncTask;
 import android.os.Build;
 
 import com.example.r_upgrade.RUpgradeFileProvider;
-import com.example.r_upgrade.common.HotUpgradeManager;
-import com.example.r_upgrade.common.IncrementUpgradeManager;
+import com.example.r_upgrade.common.install.BaseInstallFactory;
+import com.example.r_upgrade.common.install.SilentInstallFactory;
+import com.example.r_upgrade.common.manager.HotUpgradeManager;
+import com.example.r_upgrade.common.manager.IncrementUpgradeManager;
 import com.example.r_upgrade.common.RUpgradeLogger;
 import com.example.r_upgrade.common.UpgradeSQLite;
 
 import java.io.File;
 import java.lang.ref.WeakReference;
-import java.util.Map;
 
 import io.flutter.plugin.common.MethodChannel;
 
@@ -25,12 +25,14 @@ public class GenerateAndInstallByPathAsyncTask extends AsyncTask<String, Integer
     MethodChannel.Result result;
     final String path;
     final int upgradeFlavor;
+    private BaseInstallFactory installFactory;
 
-    public GenerateAndInstallByPathAsyncTask(Context context,String path,int upgradeFlavor, MethodChannel.Result result) {
+    public GenerateAndInstallByPathAsyncTask(Context context, String path, int upgradeFlavor, MethodChannel.Result result, BaseInstallFactory installFactory) {
         this.contextWrapper = new WeakReference<Context>(context);
         this.path = path;
         this.upgradeFlavor = upgradeFlavor;
         this.result = result;
+        this.installFactory = installFactory;
     }
 
     @Override
@@ -43,15 +45,21 @@ public class GenerateAndInstallByPathAsyncTask extends AsyncTask<String, Integer
             } else {
                 uri = Uri.fromFile(file);
             }
+            if(installFactory != null && installFactory instanceof SilentInstallFactory){
+                uri= Uri.parse(file.getPath());
+            }
             if (upgradeFlavor == UpgradeSQLite.UPGRADE_FLAVOR_INCREMENT) {
                 String newPath = new IncrementUpgradeManager(contextWrapper.get()).mixinAndGetNewApk(path);
-                RUpgradeLogger.get().d(TAG,"合成成功，新的安装包路径："+newPath);
+                RUpgradeLogger.get().d(TAG, "合成成功，新的安装包路径：" + newPath);
                 if (newPath == null) return null;
                 file = new File(newPath);
                 if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
                     uri = RUpgradeFileProvider.getUriForFile(contextWrapper.get(), contextWrapper.get().getApplicationInfo().packageName + ".fileProvider", file);
                 } else {
                     uri = Uri.fromFile(file);
+                }
+                if(installFactory != null && installFactory instanceof SilentInstallFactory){
+                    uri= Uri.parse(file.getPath());
                 }
             } else if (upgradeFlavor == UpgradeSQLite.UPGRADE_FLAVOR_HOT_UPDATE) {
                 boolean isSuccess = new HotUpgradeManager(contextWrapper.get()).hotUpgrade(uri);
@@ -96,20 +104,7 @@ public class GenerateAndInstallByPathAsyncTask extends AsyncTask<String, Integer
 
     //安装apk
     private boolean installApk(Uri uri) {
-        Intent install = new Intent(Intent.ACTION_VIEW);
-        if (uri != null) {
-            RUpgradeLogger.get().d(TAG, uri.toString());
-            install.setDataAndType(uri, "application/vnd.android.package-archive");
-            install.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
-                install.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
-            } else {
-                install.addCategory(Intent.CATEGORY_DEFAULT);
-            }
-            contextWrapper.get().startActivity(install);
-            return true;
-        } else {
-            return false;
-        }
+        if (installFactory == null) return true;
+        return installFactory.install(uri);
     }
 }