+++ android/src/main/java/com/example/r_upgrade/method/IRUpgradeMethodHandler.java
@@ -1,8 +1,7 @@
 package com.example.r_upgrade.method;
 
 
-import com.example.r_upgrade.common.HotUpgradeManager;
-import com.example.r_upgrade.common.UpgradeManager;
+import com.example.r_upgrade.common.manager.UpgradeManager;
 
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;