+++ android/src/main/java/com/example/r_upgrade/method/RUpgradeMethodCallHandler.java
@@ -2,14 +2,13 @@ package com.example.r_upgrade.method;
 
 import androidx.annotation.NonNull;
 
-import com.example.r_upgrade.common.HotUpgradeManager;
-import com.example.r_upgrade.common.UpgradeManager;
+import com.example.r_upgrade.common.manager.UpgradeManager;
 
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
 
 public class RUpgradeMethodCallHandler implements MethodChannel.MethodCallHandler {
-    private UpgradeManager upgradeManager;
+    private final UpgradeManager upgradeManager;
 
     public RUpgradeMethodCallHandler(UpgradeManager upgradeManager) {
         this.upgradeManager = upgradeManager;