+++ example/lib/main.dart
@@ -79,8 +79,7 @@ class MyApp extends StatefulWidget {
 
 class _MyAppState extends State<MyApp> {
   int? id;
-  bool? isAutoRequestInstall = false;
-
+  RUpgradeInstallType installType = RUpgradeInstallType.normal;
   UpgradeMethod? upgradeMethod;
 
   String? iosVersion = "";
@@ -145,6 +144,51 @@ class _MyAppState extends State<MyApp> {
         children: <Widget>[
           _buildDownloadWindow(),
           Divider(),
+          ListTile(
+            title: Text(
+              S.of(context).Install_Related,
+              style: Theme.of(context).textTheme.headline6!.copyWith(
+                    fontWeight: FontWeight.w600,
+                  ),
+            ),
+          ),
+          ListTile(
+            title: Text(
+              S.of(context).Install_Type,
+            ),
+            trailing: DropdownButton<RUpgradeInstallType>(
+              items: [
+                DropdownMenuItem(
+                  child: Text(S.of(context).Install_Type_Normal),
+                  value: RUpgradeInstallType.normal,
+                ),
+                DropdownMenuItem(
+                  child: Text(S.of(context).Install_Type_Silent),
+                  value: RUpgradeInstallType.silent,
+                ),
+                DropdownMenuItem(
+                  child: Text(S.of(context).Install_Type_None),
+                  value: RUpgradeInstallType.none,
+                ),
+              ],
+              value: installType,
+              onChanged: (value) {
+                if (value == null) return;
+                setState(() {
+                  installType = value;
+                });
+              },
+            ),
+          ),
+          if (installType == RUpgradeInstallType.silent)
+            ListTile(
+              title: Text(
+                S.of(context).Install_Type_Silent_Tip,
+                style: Theme.of(context).textTheme.subtitle2?.copyWith(
+                      color: Theme.of(context).errorColor,
+                    ),
+              ),
+            ),
           ListTile(
             title: Text(
               S.of(context).Update_the_related,
@@ -220,9 +264,9 @@ class _MyAppState extends State<MyApp> {
               id = await RUpgrade.upgrade(
 //                "http://192.168.1.105:8888/files/static/kuan.apk",
 //                  'http://dl-cdn.coolapkmarket.com/down/apk_file/2020/0308/Coolapk-v10.0.3-2003081-coolapk-app-release.apk?_upt=b210caeb1585012557',
-                  'https://mydata-1252536312.cos.ap-guangzhou.myqcloud.com/r_upgrade.apk',
+                  'https://dl-tc.coolapkmarket.com/down/apk_file/2022/1031/Coolapk-12.5.1-2210311-coolapk-app-sign.apk?t=1675873958&sign=c2eed2272fe3ac07435a555dbbd5bfe4',
                   fileName: 'r_upgrade.apk',
-                  isAutoRequestInstall: isAutoRequestInstall!,
+                  installType: installType!,
                   notificationStyle: NotificationStyle.speechAndPlanTime,
                   useDownloadManager: false);
               upgradeMethod = UpgradeMethod.all;
@@ -249,7 +293,8 @@ class _MyAppState extends State<MyApp> {
               final status = await RUpgrade.getDownloadStatus(id!);
 
               if (status == DownloadStatus.STATUS_SUCCESSFUL) {
-                bool? isSuccess = await RUpgrade.install(id!);
+                bool? isSuccess =
+                    await RUpgrade.install(id!, installType: installType);
                 if (isSuccess != null && isSuccess) {
                   ScaffoldMessenger.of(context).showSnackBar(SnackBar(
                       content: Text(S.of(context).The_request_is_successful)));
@@ -259,15 +304,6 @@ class _MyAppState extends State<MyApp> {
                     content: Text(S.of(context).The_current_ID_not_download)));
               }
             },
-          ),
-          CheckboxListTile(
-            value: isAutoRequestInstall,
-            onChanged: (bool? value) {
-              setState(() {
-                isAutoRequestInstall = value;
-              });
-            },
-            title: Text(S.of(context).After_download_to_install),
           ),
           ListTile(
             title: Text(S.of(context).Continue_to_update),
@@ -336,7 +372,7 @@ class _MyAppState extends State<MyApp> {
                   'https://mydata-1252536312.cos.ap-guangzhou.myqcloud.com/r_upgrade.zip',
                   fileName: 'r_upgrade.zip',
                   useDownloadManager: false,
-                  isAutoRequestInstall: false,
+                  installType: installType,
                   upgradeFlavor: RUpgradeFlavor.hotUpgrade);
               upgradeMethod = UpgradeMethod.hot;
               setState(() {});
@@ -363,7 +399,8 @@ class _MyAppState extends State<MyApp> {
               final status = await RUpgrade.getDownloadStatus(id!);
 
               if (status == DownloadStatus.STATUS_SUCCESSFUL) {
-                bool? isSuccess = await RUpgrade.install(id!);
+                bool? isSuccess =
+                    await RUpgrade.install(id!, installType: installType);
                 if (isSuccess != null && isSuccess) {
                   ScaffoldMessenger.of(context).showSnackBar(SnackBar(
                       content: Text(S
@@ -407,7 +444,7 @@ class _MyAppState extends State<MyApp> {
                 'https://mydata-1252536312.cos.ap-guangzhou.myqcloud.com/r_upgrade.patch',
                 fileName: 'r_upgrade.patch',
                 useDownloadManager: false,
-                isAutoRequestInstall: false,
+                installType: installType,
                 upgradeFlavor: RUpgradeFlavor.incrementUpgrade,
               );
               upgradeMethod = UpgradeMethod.increment;
@@ -436,7 +473,7 @@ class _MyAppState extends State<MyApp> {
               try {
                 final status = await RUpgrade.getDownloadStatus(id!);
                 if (status == DownloadStatus.STATUS_SUCCESSFUL) {
-                  await RUpgrade.install(id!);
+                  await RUpgrade.install(id!, installType: installType);
                 } else {
                   ScaffoldMessenger.of(context).showSnackBar(SnackBar(
                       content: